+++ creme/linear_model/__init__.py
@@ -5,22 +5,10 @@ the :mod:`creme.optim` module.
 from .fm import FMRegressor
 from .linear_regression import LinearRegression
 from .logistic_regression import LogisticRegression
-from .pa import PassiveAggressiveClassifier
-from .pa import PassiveAggressiveRegressor
-from .pa import PassiveAggressiveIClassifier
-from .pa import PassiveAggressiveIRegressor
-from .pa import PassiveAggressiveIIClassifier
-from .pa import PassiveAggressiveIIRegressor
 
 
 __all__ = [
     'FMRegressor',
     'LinearRegression',
-    'LogisticRegression',
-    'PassiveAggressiveClassifier',
-    'PassiveAggressiveRegressor',
-    'PassiveAggressiveIClassifier',
-    'PassiveAggressiveIRegressor',
-    'PassiveAggressiveIIClassifier',
-    'PassiveAggressiveIIRegressor'
+    'LogisticRegression'
 ]