+++ creme/linear_model/logistic_regression.py
@@ -2,8 +2,7 @@ import collections
 
 from .. import base
 from .. import optim
-
-from . import util
+from .. import utils
 
 
 __all__ = ['LogisticRegression']
@@ -55,10 +54,10 @@ class LogisticRegression(base.BinaryClassifier):
         self.weights = collections.defaultdict(float)
 
     def _predict_proba_one_with_weights(self, x, w):
-        return util.sigmoid(util.dot(x, w))
+        return utils.sigmoid(utils.dot(x, w))
 
-    def _calc_gradient(self, y_true, y_pred, x, w):
-        loss_gradient = self.loss.gradient(y_true, y_pred)
+    def _calc_gradient(self, y_true, y_pred, loss, x, w):
+        loss_gradient = loss.gradient(y_true, y_pred)
         return {i: xi * loss_gradient + self.l2 * w.get(i, 0) for i, xi in x.items()}
 
     def fit_one(self, x, y):
@@ -68,6 +67,7 @@ class LogisticRegression(base.BinaryClassifier):
             x=x,
             y=y,
             w=self.weights,
+            loss=self.loss,
             f_pred=self._predict_proba_one_with_weights,
             f_grad=self._calc_gradient
         )