+++ creme/optim/ada_grad.py
@@ -42,7 +42,7 @@ class AdaGrad(base.Optimizer):
     def __init__(self, lr=0.1, eps=1e-8):
         super().__init__(lr)
         self.eps = eps
-        self.g2 = collections.defaultdict(lambda: 0.)
+        self.g2 = collections.defaultdict(float)
 
     def update_weights_with_gradient(self, w, g):