+++ creme/optim/adam.py
@@ -44,16 +44,16 @@ class Adam(base.Optimizer):
         self.beta_1 = beta_1
         self.beta_2 = beta_2
         self.eps = eps
-        self.m = collections.defaultdict(lambda: 0.)
-        self.v = collections.defaultdict(lambda: 0.)
+        self.m = collections.defaultdict(float)
+        self.v = collections.defaultdict(float)
 
     def update_weights_with_gradient(self, w, g):
 
         for i, gi in g.items():
             self.m[i] = self.beta_1 * self.m[i] + (1 - self.beta_1) * gi
             self.v[i] = self.beta_2 * self.v[i] + (1 - self.beta_2) * gi ** 2
-            m = self.m[i] / (1 - self.beta_1 ** (self.t + 1))
-            v = self.v[i] / (1 - self.beta_2 ** (self.t + 1))
+            m = self.m[i] / (1 - self.beta_1 ** (self.n_iterations + 1))
+            v = self.v[i] / (1 - self.beta_2 ** (self.n_iterations + 1))
             w[i] -= self.learning_rate * m / (v ** 0.5 + self.eps)
 
         return w