+++ creme/optim/base.py
@@ -8,16 +8,16 @@ class Optimizer(abc.ABC):
 
     def __init__(self, lr: typing.Union[lr_schedule.LRScheduler, float]):
         self.lr = lr_schedule.ConstantLR(lr) if isinstance(lr, float) else lr
-        self.t = 0
+        self.n_iterations = 0
 
     @property
     def learning_rate(self) -> float:
-        return self.lr.get(self.t)
+        return self.lr.get(self.n_iterations)
 
     def update_weights_with_gradient(self, w: dict, g: dict) -> dict:
         raise NotImplementedError
 
-    def update_weights(self, x, y, w, f_pred, f_grad):
+    def update_weights(self, x, y, w, loss, f_pred, f_grad):
         """
 
         This method gives more control over the gradient computation than
@@ -32,11 +32,12 @@ class Optimizer(abc.ABC):
         y_pred = f_pred(x, w)
 
         # Compute the gradient
-        gradient = f_grad(y, y_pred, x, w)
+        gradient = f_grad(y_true=y, y_pred=y_pred, loss=loss, x=x, w=w)
 
         # Update the weights
         w = self.update_weights_with_gradient(w, gradient)
 
-        self.t += 1
+        # Update the iteration counter
+        self.n_iterations += 1
 
         return w, y_pred